requirements.txt
aiohappyeyeballs        = 2.4.0
aiohttp                 = 3.10.5
aiosignal               = 1.3.1
annotated-types         = 0.7.0
anyio                   = 4.4.0
attrs                   = 24.2.0
blinker                 = 1.8.2
boto3                   = 1.35.13
botocore                = 1.35.13
camelot-py              = 0.9.0
certifi                 = 2024.8.30
cffi                    = 1.17.1
chardet                 = 5.2.0
charset-normalizer      = 3.3.2
click                   = 8.1.7
cohere                  = 5.9.1
colorama                = 0.4.6
colorlog                = 6.8.2
contourpy               = 1.3.0
cryptography            = 43.0.1
cycler                  = 0.12.1
dataclasses-json        = 0.6.7
distro                  = 1.9.0
et-xmlfile              = 1.1.0
faiss-cpu               = 1.8.0.post1
fastavro                = 1.9.7
filelock                = 3.15.4
Flask                   = 3.0.3
Flask-Cors              = 5.0.0
fonttools               = 4.53.1
frozenlist              = 1.4.1
fsspec                  = 2024.9.0
ftfy                    = 6.2.3
greenlet                = 3.0.3
h11                     = 0.14.0
httpcore                = 1.0.5
httpx                   = 0.27.2
httpx-sse               = 0.4.0
huggingface-hub         = 0.24.6
idna                    = 3.8
itsdangerous            = 2.2.0
Jinja2                  = 3.1.4
jiter                   = 0.5.0
jmespath                = 1.0.1
joblib                  = 1.4.2
jsonpatch               = 1.33
jsonpointer             = 3.0.0
kiwisolver              = 1.4.7
langchain               = 0.2.16
langchain-community     = 0.2.16
langchain-core          = 0.2.38
langchain-text-splitters= 0.2.4
langsmith               = 0.1.115
MarkupSafe              = 2.1.5
marshmallow             = 3.22.0
matplotlib              = 3.9.2
mpmath                  = 1.3.0
multidict               = 6.0.5
mypy-extensions         = 1.0.0
networkx                = 3.3
nltk                    = 3.9.1
numpy                   = 1.26.4
open_clip_torch         = 2.26.1
openai                  = 1.44.0
opencv-python           = 4.10.0.84
openpyxl                = 3.1.5
orjson                  = 3.10.7
packaging               = 24.1
pandas                  = 2.2.2
parameterized           = 0.9.0
pdf2image               = 1.17.0
pdfminer.six            = 20240706
pillow                  = 10.4.0
pip                     = 24.2
psycopg2-binary         = 2.9.9
pycparser               = 2.22
pydantic                = 2.9.0
pydantic_core           = 2.23.2
PyMuPDF                 = 1.24.10
PyMuPDFb                = 1.24.10
pyparsing               = 3.1.4
PyPDF2                  = 3.0.1
pytesseract             = 0.3.13
python-dateutil         = 2.9.0.post0
pytz                    = 2024.1
PyYAML                  = 6.0.2
regex                   = 2023.12.25
requests                = 2.32.3
requests-mock           = 1.12.1
s3transfer              = 0.10.2
safetensors             = 0.4.5
scikit-learn            = 1.5.1
scipy                   = 1.14.1
semantic-chunkers       = 0.0.9
semantic-router         = 0.0.65
sentence-transformers   = 3.0.1
setuptools              = 72.1.0
six                     = 1.16.0
sniffio                 = 1.3.1
SQLAlchemy              = 2.0.34
sympy                   = 1.13.2
tenacity                = 8.5.0
threadpoolctl           = 3.5.0
tiktoken                = 0.7.0
timm                    = 1.0.9
tokenizers              = 0.19.1
torch                   = 2.4.1
torchvision             = 0.19.1
tqdm                    = 4.66.5
transformers            = 4.44.2
types-requests          = 2.32.0.20240905
typing_extensions       = 4.12.2
typing-inspect          = 0.9.0
tzdata                  = 2024.1
urllib3                 = 2.2.2
wcwidth                 = 0.2.13
Werkzeug                = 3.0.4
wheel                   = 0.43.0
yarl                    = 1.9.11